+++ assets/js/app/app.component.js
@@ -1,4 +1,4 @@
-System.register("app.component", ['angular2/core'], function(exports_1, context_1) {
+System.register("app.component", ['@angular/core'], function(exports_1, context_1) {
     "use strict";
     var __moduleName = context_1 && context_1.id;
     var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
