+++ assets/js/app/chat-detail.component.js
