+++ assets/js/app/chat-detail.component.ts
