+++ package.json
@@ -21,13 +21,21 @@
     "rc": "~0.5.0",
     "sails": "~0.11.4",
     "sails-disk": "~0.10.0",
-    "angular2": "2.0.0-beta.3",
-    "systemjs": "0.19.6",
+    "rxjs": "5.0.0-beta.6",
+    "@angular/common": "2.0.0-rc.4",
+    "@angular/compiler": "2.0.0-rc.4",
+    "@angular/core": "2.0.0-rc.4",
+    "@angular/http": "2.0.0-rc.4",
+    "@angular/platform-browser": "2.0.0-rc.4",
+    "@angular/platform-browser-dynamic": "2.0.0-rc.4",
+    "@angular/router": "3.0.0-beta.1",
+    "@angular/router-deprecated": "2.0.0-rc.2",
+    "@angular/upgrade": "2.0.0-rc.4",
+    "systemjs": "^0.19.8",
     "es6-promise": "^3.0.2",
-    "es6-shim": "^0.33.3",
+    "es6-shim": "^0.35.0",
     "reflect-metadata": "0.1.2",
-    "rxjs": "5.0.0-beta.0",
-    "zone.js": "0.5.11",
+    "zone.js": "^0.6.12",
     "sails-hook-typescript": "^1.0.0"
   },
   "devDependencies": {
