+++ tasks/config/copy.js
@@ -36,12 +36,11 @@ module.exports = function(grunt) {
 			files: [{
 				expand: true,
 				src: [
-					'./node_modules/es6-shim/es6-shim.min.js',
-					'./node_modules/systemjs/dist/system-polyfills.js',
-					'./node_modules/angular2/bundles/angular2-polyfills.js',
-					'./node_modules/systemjs/dist/system.src.js',
-					'./node_modules/rxjs/bundles/Rx.js',
-					'./node_modules/angular2/bundles/angular2.dev.js'
+					// './node_modules/es6-shim/es6-shim.min.js',
+					// './node_modules/systemjs/dist/system-polyfills.js',
+					// './node_modules/systemjs/dist/system.src.js',
+					// './node_modules/rxjs/bundles/Rx.js',
+					'./node_modules/**/*'
 				],
 				dest: '.tmp/public/js'
 			}]
