+++ views/layout.ejs
@@ -90,26 +90,77 @@
     <!--SCRIPTS-->
     <script src="/js/dependencies/sails.io.js"></script>
     <script src="/js/node_modules/es6-shim/es6-shim.min.js"></script>
+    <script src="/js/node_modules/reflect-metadata/Reflect.js"></script>
+    <script src="/js/node_modules/zone.js/dist/zone.min.js"></script>
     <script src="/js/node_modules/systemjs/dist/system-polyfills.js"></script>
-    <script src="/js/node_modules/angular2/bundles/angular2-polyfills.js"></script>
     <script src="/js/node_modules/systemjs/dist/system.src.js"></script>
-    <script src="/js/node_modules/rxjs/bundles/Rx.js"></script>
-    <script src="/js/node_modules/angular2/bundles/angular2.dev.js"></script>
-    <script src="/js/app.component.js"></script>
-    <script src="/js/chat-detail.component.js"></script>
-    <script src="/js/chat.js"></script>
-    <script src="/js/main.js"></script>
+    <script src="/js/node_modules/rxjs/bundles/Rx.min.js"></script>
+    <script src="/js/app/app.component.js"></script>
+    <script src="/js/app/chat-detail.component.js"></script>
+    <script src="/js/app/chat.js"></script>
+    <script src="/js/app/main.js"></script>
     <!--SCRIPTS END-->
     
     <script src="http://localhost:35729/livereload.js"></script>
     <script>
       System.config({
-        packages: {        
-          app: {
-            format: 'register',
-            defaultExtension: 'js'
+          packages: {
+              app: {
+                  format: 'register',
+                  defaultExtension: 'js',
+                  main: 'main.js'
+              },
+              '@angular/core': {
+                  defaultExtension: 'js',
+                  main: 'index.js'
+              },
+
+              '@angular/common': {
+                  defaultExtension: 'js',
+                  main: 'index.js'
+              },
+
+              '@angular/compiler': {
+                  defaultExtension: 'js',
+                  main: 'index.js'
+              },
+
+              '@angular/router': {
+                  defaultExtension: 'js',
+                  main: 'index.js'
+              },
+
+              '@angular/router-deprecated': {
+                  defaultExtension: 'js',
+                  main: 'index.js'
+              },
+
+              '@angular/http': {
+                  defaultExtension: 'js',
+                  main: 'index.js'
+              },
+
+              '@angular/platform-browser': {
+                  defaultExtension: 'js',
+                  main: 'index.js'
+              },
+
+              '@angular/platform-browser-dynamic': {
+                  defaultExtension: 'js',
+                  main: 'index.js'
+              }
+          },
+          map: {
+              'app': 'js/app',
+              '@angular/core': 'js/node_modules/@angular/core',
+              '@angular/common': 'js/node_modules/@angular/common',
+              '@angular/compiler': 'js/node_modules/@angular/compiler',
+              '@angular/router': 'js/node_modules/@angular/router',
+              '@angular/router-deprecated': 'js/node_modules/@angular/router-deprecated',
+              '@angular/http': 'js/node_moduleslib/@angular/http',
+              '@angular/platform-browser': 'js/node_modules/@angular/platform-browser',
+              '@angular/platform-browser-dynamic': 'js/node_modules/@angular/platform-browser-dynamic'
           }
-        }
       });
       System.import('main')
             .then(null, console.error.bind(console));
